In decodable format, but problem making the JSON global

+++ Clima/WeatherDataModel.swift
@@ -8,12 +8,13 @@ import UIKit
 
 class WeatherDataModel {
 
-    var temperature : Int = 0
+    var temperature : Double = 0.0
     var condition : Int = 0
     var city : String = ""
     var weatherIconName : String = ""
     
     
+
